feat-fix: allow unknoen keys in base schema validation
