+++ src/cli/repl/server/messages/analysis.ts
@@ -1,7 +1,6 @@
 import { FlowrBaseMessage, RequestMessageDefinition } from './messages'
 import { LAST_PER_FILE_STEP, StepResults } from '../../../../core'
 import Joi from 'joi'
-import { FlowRServerConnection } from '../connection'
 
 export interface FileAnalysisRequestMessage extends FlowrBaseMessage {
 	type:      'request-file-analysis',
@@ -11,18 +10,14 @@ export interface FileAnalysisRequestMessage extends FlowrBaseMessage {
 }
 
 
-export const requestAnalysisMessage: RequestMessageDefinition<FileAnalysisRequestMessage, [FlowRServerConnection]> = {
+export const requestAnalysisMessage: RequestMessageDefinition<FileAnalysisRequestMessage> = {
 	type:   'request-file-analysis',
 	schema: Joi.object({
 		type:      Joi.string().valid('request-file-analysis').required(),
 		filetoken: Joi.string().required(),
 		filename:  Joi.string().required(),
 		content:   Joi.string().required()
-	}),
-
-	handle(message: FileAnalysisRequestMessage, information: [FlowRServerConnection]) {
-		console.log('received file analysis request')
-	}
+	})
 }
