--- src/cli/repl/server/messages/index.ts
@@ -1,5 +1,0 @@
-export * from './error'
-export * from './hello'
-export * from './slice'
-export * from './analysis'
-export * from './messages'
